renamed base file for Environment class
