+++ src/rl_framework/environment/__init__.py
@@ -1,2 +1,2 @@
-from .environment import Environment
+from .base_environment import Environment
 from .gym_environment import GymEnvironmentWrapper
